+++ src/test/java/ssammudan/cotree/integration/SpringBootTestSupporter.java
@@ -14,7 +14,7 @@ import jakarta.persistence.EntityManager;
 import software.amazon.awssdk.services.s3.S3Client;
 import ssammudan.cotree.domain.community.service.CommunityService;
 import ssammudan.cotree.domain.education.techbook.service.TechBookService;
-import ssammudan.cotree.domain.review.review.service.TechEducationReviewService;
+import ssammudan.cotree.domain.review.service.TechEducationReviewService;
 import ssammudan.cotree.infra.s3.S3Uploader;
 import ssammudan.cotree.integration.factory.CommunityDataFactory;
 import ssammudan.cotree.integration.factory.MemberDataFactory;
